Refine blend core crates
