+++ nomos-blend/scheduling/Cargo.toml
@@ -12,7 +12,6 @@ async-trait         = { default-features = false, version = "0.1" }
 derivative          = "2"
 fork_stream         = { workspace = true }
 futures             = { default-features = false, version = "0.3" }
-groth16             = { default-features = false, workspace = true }
 hex                 = { default-features = false, version = "0.4" }
 multiaddr           = { default-features = false, version = "0.18" }
 nomos-blend-message = { workspace = true }
