+++ nomos-services/blend/src/core/settings.rs
@@ -2,7 +2,7 @@ use std::num::NonZeroU64;
 
 use futures::{Stream, StreamExt as _};
 use nomos_blend_scheduling::{
-    membership::Membership, message_blend::CryptographicProcessorSettings,
+    membership::Membership, message_blend::SessionCryptographicProcessorSettings,
     message_scheduler::session_info::SessionInfo, session::SessionEvent,
 };
 use nomos_utils::math::NonNegativeF64;
@@ -13,7 +13,7 @@ use crate::settings::TimingSettings;
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct BlendConfig<BackendSettings> {
     pub backend: BackendSettings,
-    pub crypto: CryptographicProcessorSettings,
+    pub crypto: SessionCryptographicProcessorSettings,
     pub scheduler: SchedulerSettingsExt,
     pub time: TimingSettings,
     pub minimum_network_size: NonZeroU64,
@@ -40,7 +40,7 @@ pub struct CoverTrafficSettingsExt {
 impl CoverTrafficSettingsExt {
     fn session_quota(
         &self,
-        crypto: &CryptographicProcessorSettings,
+        crypto: &SessionCryptographicProcessorSettings,
         timings: &TimingSettings,
         membership_size: usize,
     ) -> u64 {
@@ -196,7 +196,7 @@ mod tests {
     fn settings(message_frequency_per_round: f64, rounds_per_session: u64) -> BlendConfig<()> {
         BlendConfig {
             backend: (),
-            crypto: CryptographicProcessorSettings {
+            crypto: SessionCryptographicProcessorSettings {
                 non_ephemeral_signing_key: Ed25519PrivateKey::generate(),
                 num_blend_layers: 1,
             },
