+++ nomos-services/blend/src/edge/settings.rs
@@ -1,6 +1,6 @@
 use core::num::NonZeroU64;
 
-use nomos_blend_scheduling::message_blend::CryptographicProcessorSettings;
+use nomos_blend_scheduling::message_blend::SessionCryptographicProcessorSettings;
 use serde::{Deserialize, Serialize};
 
 use crate::settings::TimingSettings;
@@ -8,7 +8,7 @@ use crate::settings::TimingSettings;
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct BlendConfig<BackendSettings> {
     pub backend: BackendSettings,
-    pub crypto: CryptographicProcessorSettings,
+    pub crypto: SessionCryptographicProcessorSettings,
     pub time: TimingSettings,
     pub minimum_network_size: NonZeroU64,
 }
